requirements.txt
# requirements.txt
streamlit
spotipy
spacy
nltk
python-dotenv
rapidfuzz
requests
sentence-transformers